requirements.txt
pbr >= 5.8.1 
pychroot >= 0.10.4
pyyaml >= 5.0